# search: Add a minimal search option