@gasche 2015-03-12
After a quick review, the patch is obviously correct.