@letouzey 2015-03-12
Hi,

I support this pull request, in the long run we will integrate this kind of
queries in the ide<->coq protocol but for now the feature is nice and
probably easy to have in coqide too.

P.



2015-03-12 16:51 GMT+01:00 gasche <notifications@github.com>:

> After a quick review, the patch is obviously correct.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/coq/coq/pull/56#issuecomment-78507992>.
>