@herbelin 2015-03-18
OK to me too, but maybe could you justify why controlling the format with a global option and not locally by using a modifier to "Search", in the style

Search _ (short format).
SearchAbout query ... query (short format) inside mod ... mod.
SearchAbout query ... query (short format) outside mod ... mod.

maybe actually suggesting an alternative syntax

SearchAbout query ... query (short format, inside mod ... mod).
SearchAbout query ... query (short format, outside mod ... mod).

reminiscent of the syntax for Notation modifiers.