@herbelin 2015-03-18
(And similarly, of course, for SearchPattern, SearchRewrite, SearchHead, and the Search shortening of SearchAbout.)