@cpitclaudel 2015-03-18
Mostly for consistency with the `Set Printing ...` options. Also, I didn't feel comfortable touching too much of the code, and adding one flag seemed simpler than modifying the existing commands :)