@herbelin 2015-03-18
Btw, Pierre, what interface to (short) Search do you have in mind when called from CoqIDE?