@letouzey 2015-03-18
This would be nice for completion of identifiers via a popup menu that
displays the type of ids when cycling on them (that's an example, Clément
will make a demo!).

The idea is that it is better to ask for names only to display the menu
quickly without to much parsing, and then ask for types on the fly when
cycling in the menu.

P.




2015-03-18 22:09 GMT+01:00 Hugo Herbelin <notifications@github.com>:

> Btw, Pierre, what interface to (short) Search do you have in mind when
> called from CoqIDE?
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/coq/coq/pull/56#issuecomment-83177568>.
>