@herbelin 2015-03-19
I looked at your package, which looks very very nice. Then, your motivation is efficiency and, implicitly, stability of the output so that it can be mechanically processed. So maybe there is no need to care about a proper user-level syntax for it, and a flag is well enough.

I don't see any reason to have the flag in constrextern.ml and its declaration in vernacentries.ml. The appropriate location for both the flag and its declaration seems search.ml to me. I would also suggest you add a comment in the short output mode telling that the output is intended to be mechanically processed so that noone has accidentally the idea to condense the output as in "id1, id2, ..." rather than one name per line.

I don't find "Search Minimal" to be a very discriminating name. Maybe "Search Output Name Only", or "Search Short Output".