@cpitclaudel 2015-03-19
Noted for the low-specificity name; I don't have any particular preference :) The comment seems a good idea too. I'll add it to the source. Your point about the lack of a need for a user-level syntax also makes a lot of sense.

I wasn't sure where to add the options, so I drew inspiration from the `Set Printing` ones. I'll be happy to move them if you think it's better to put them in search.ml.