@herbelin 2015-03-23
> If you think it's better to put them in search.ml.

Yes, this would avoid exporting an extra mutable variable and increase modularity.

> I don't have any particular preference

So, let's say "Search Output Name Only", with message "The output-name-only search mode is", if you are ok.