@cpitclaudel 2015-03-24
Sounds good to me! I've updated the PR; I'll squash the commits when we reach the final iteration :)