@cpitclaudel 2015-04-21
I just moved to 8.5 to do more testing, and I'm not too sure about merging this PR any more. The time it takes to print things seems to be quadratic in 8.5, which makes search commands rather slow when printing a large number of identifiers. Perhaps I should open a separate issue about this.