@cpitclaudel 2015-04-27
I've added documentation, and Guillaume fixed (98a710caf5e907344329ee9e9f7b5fd87c50836f) the regression that I mentioned above, so this is good to go as far as I'm concerned!