@cpitclaudel 2015-07-25
Ping — any reason to not merge this?