@mattam82 2015-07-27
Thanks! We discussed it at the Coq WG and accepted it.