@cpitclaudel 2016-01-02
Just to clarify: this got merged into trunk in July, so it won't be in 8.5, correct?